Fix: int cannot compare to size_t

+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: hbui <hbui@student.hive.fi>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/01 18:02:03 by hbui              #+#    #+#             */
-/*   Updated: 2021/12/02 08:50:11 by hbui             ###   ########.fr       */
+/*   Updated: 2021/12/07 22:25:05 by hbui             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*str_s;
 	unsigned char	unsigned_c;
